requirements.txt
python>=3.10

torch>=2.0
torchvision>=0.15

numpy
matplotlib
pillow
